qrconf binder  && draw text size fix
