+++ src/pages/index.astro
@@ -37,35 +37,32 @@
                     <button type="button" class="accordion accordion--open">Main Options</button>
                     <div class="panel panel--open">
                         <label for="form-data">Data</label>
-                        <textarea data-node="data" node-change-event="oninput" id="form-data" value="https://qr-code-styling.com">https://qr-code-styling.com</textarea>
+                        <textarea data-node="data" node-change-event="oninput" id="form-data">https://qr.yesil.dev</textarea>
                         <label for="form-image-file">Image File</label>
                         <div class="buttons-container">
-                            <input data-node="image" node-data-field="files" id="form-image-file" type="file" accept="image/*" />
+                            <input data-node="image" node-data-field="files" id="form-image-file" type="file" accept="image/*"  value="linux"/>
                             <button type="button" id="button-cancel">Cancel</button>
                         </div>
                         
-                        <label for="form-width">Width</label>
-                        <div>
+                        <div class="group">
+                            <label for="form-width">Width</label>
                             <input data-node="width" id="form-width" type="number" min="100" max="10000" value="600"/>
-                        </div>
-                        <label for="form-height">Height</label>
-                        <div>
+                            <label for="form-height">Height</label>
                             <input data-node="height" id="form-height" type="number" min="100" max="10000" value="600"/>
-                        </div>
-                        <label for="form-height">Margin</label>
-                        <div>
+                            <label for="form-height">Margin</label>
                             <input data-node="margin" id="form-margin" type="number" min="0" max="10000" value="20"/>
                         </div>
-               
+
                         <label for="form-text" title="You can use Markdown format">Text</label>
-                        <textarea data-node="text" title="You can use Markdown format" node-change-event="oninput" id="form-text" value="">text</textarea>
-                        <label for="form-text-pos">Text Pos</label>
-                        <div>
+                        <textarea data-node="text" title="You can use Markdown format" node-change-event="oninput" id="form-text" >Title text</textarea>
+                        <div class="group">
+                            <label for="form-text-pos">Text Pos</label>
                             <select data-node="textpos" id="form-text-pos">
                                 <option value="right">Right</option>
                                 <option value="bottom" selected>Bottom</option>
-                             
                             </select>
+                            <label for="form-font-factor">Font Size Factor</label>
+                            <input data-node="textFontFactor" id="form-font-factor" type="number" min="0.1" step="0.1" value="1.0"/>
                         </div>
                     </div>
                     <button type="button" class="accordion">Dots Options</button>
