+++ src/scripts/form-util.js
@@ -1,11 +1,3 @@
-export async function toBase64(file) {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onload = () => resolve(reader.result);
-    reader.onerror = reject;
-  });
-}
 
 export async function getFormData(inputList) {
   const fdata = {};
