+++ src/scripts/tools.js
@@ -62,3 +62,11 @@ export function getSrcFromFile(file, callback) {
 
     reader.readAsDataURL(file);
 }
+export async function toBase64(file) {
+    return new Promise((resolve, reject) => {
+      const reader = new FileReader();
+      reader.readAsDataURL(file);
+      reader.onload = () => resolve(reader.result);
+      reader.onerror = reject;
+    });
+  }
